# services/graph_harness_api/openapi.json
{
  "components": {
    "schemas": {
...
      "ContinuousLearningRunRequest": {
        "description": "Request payload for one continuous-learning cycle.",
        "properties": {
          "max_depth": {
            "default": 2,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Max Depth",
            "type": "integer"
          },
          "max_new_proposals": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max New Proposals",
            "type": "integer"
          },
          "max_next_questions": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Next Questions",
            "type": "integer"
          },
          "model_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "relation_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 200,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation Types"
          },
          "run_budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HarnessRunBudget"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedule_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "seed_entity_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Entity Ids"
          },
          "source_type": {
            "default": "pubmed",
            "maxLength": 64,
            "minLength": 1,
            "title": "Source Type",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ContinuousLearningRunRequest",
        "type": "object"
      },
      "ContinuousLearningRunResponse": {
        "description": "Combined response for a completed continuous-learning cycle.",
        "properties": {
          "budget_status": {
            "$ref": "#/components/schemas/HarnessRunBudgetStatus"
          },
          "candidate_count": {
            "title": "Candidate Count",
            "type": "integer"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ContinuousLearningCandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "delta_report": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Delta Report",
            "type": "object"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "next_questions": {
            "items": {
              "type": "string"
            },
            "title": "Next Questions",
            "type": "array"
          },
          "proposal_count": {
            "title": "Proposal Count",
            "type": "integer"
          },
          "run": {
            "$ref": "#/components/schemas/HarnessRunResponse"
          },
          "run_budget": {
            "$ref": "#/components/schemas/HarnessRunBudget"
          }
        },
        "required": [
          "run",
          "candidates",
          "candidate_count",
          "proposal_count",
          "next_questions",
          "delta_report",
          "errors",
          "run_budget",
          "budget_status"
        ],
        "title": "ContinuousLearningRunResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HarnessRunResponse": {
        "description": "Serialized harness run record.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "graph_service_status": {
            "title": "Graph Service Status",
            "type": "string"
          },
          "graph_service_version": {
            "title": "Graph Service Version",
            "type": "string"
          },
          "harness_id": {
            "title": "Harness Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Input Payload",
            "type": "object"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "harness_id",
          "title",
          "status",
          "input_payload",
          "graph_service_status",
          "graph_service_version",
          "created_at",
          "updated_at"
        ],
        "title": "HarnessRunResponse",
        "type": "object"
      },
...
      "HarnessTemplateListResponse": {
        "description": "List response for harness template discovery.",
        "properties": {
          "harnesses": {
            "items": {
              "$ref": "#/components/schemas/HarnessTemplateResponse"
            },
            "title": "Harnesses",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "harnesses",
          "total"
        ],
        "title": "HarnessTemplateListResponse",
        "type": "object"
      },
      "HarnessTemplateResponse": {
        "description": "Serialized harness template.",
        "properties": {
          "default_run_budget": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JSONValue-Output"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Run Budget"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outputs": {
            "items": {
              "type": "string"
            },
            "title": "Outputs",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tool_groups": {
            "items": {
              "type": "string"
            },
            "title": "Tool Groups",
            "type": "array"
          }
        },
        "required": [
          "id",
          "display_name",
          "summary",
          "tool_groups",
          "outputs"
        ],
        "title": "HarnessTemplateResponse",
        "type": "object"
      },
...
      "JSONValue-Output": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "object"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "array"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "SupervisorRunDetailResponse": {
        "description": "Persisted supervisor run state for typed reloads.",
        "properties": {
          "artifact_keys": {
            "$ref": "#/components/schemas/SupervisorArtifactKeysResponse"
          },
          "bootstrap": {
            "$ref": "#/components/schemas/ResearchBootstrapRunResponse"
          },
          "bootstrap_run_id": {
            "title": "Bootstrap Run Id",
            "type": "string"
          },
          "briefing_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Briefing Question"
          },
          "chat": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatMessageRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "chat_graph_write_proposal_ids": {
            "items": {
              "type": "string"
            },
            "title": "Chat Graph Write Proposal Ids",
            "type": "array"
          },
          "chat_graph_write_review_count": {
            "title": "Chat Graph Write Review Count",
            "type": "integer"
          },
          "chat_graph_write_reviews": {
            "items": {
              "$ref": "#/components/schemas/SupervisorChatGraphWriteReviewResponse"
            },
            "title": "Chat Graph Write Reviews",
            "type": "array"
          },
          "chat_graph_write_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Graph Write Run Id"
          },
          "chat_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Run Id"
          },
          "chat_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Session Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "curation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimCurationRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "curation_actions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JSONValue-Output"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curation Actions"
          },
          "curation_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curation Run Id"
          },
          "curation_source": {
            "title": "Curation Source",
            "type": "string"
          },
          "curation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curation Status"
          },
          "curation_summary": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JSONValue-Output"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curation Summary"
          },
          "latest_chat_graph_write_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupervisorChatGraphWriteReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "$ref": "#/components/schemas/HarnessRunProgressResponse"
          },
          "run": {
            "$ref": "#/components/schemas/HarnessRunResponse"
          },
          "selected_curation_proposal_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Curation Proposal Ids",
            "type": "array"
          },
          "skipped_steps": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Steps",
            "type": "array"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/SupervisorStepResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "workflow": {
            "title": "Workflow",
            "type": "string"
          }
        },
        "required": [
          "run",
          "progress",
          "workflow",
          "bootstrap",
          "chat",
          "curation",
          "artifact_keys",
          "bootstrap_run_id",
          "chat_run_id",
          "chat_session_id",
          "chat_graph_write_run_id",
          "curation_run_id",
          "briefing_question",
          "curation_source",
          "curation_status",
          "completed_at",
          "chat_graph_write_proposal_ids",
          "selected_curation_proposal_ids",
          "skipped_steps",
          "chat_graph_write_review_count",
          "latest_chat_graph_write_review",
          "chat_graph_write_reviews",
          "steps",
          "curation_summary",
          "curation_actions"
        ],
        "title": "SupervisorRunDetailResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MED13 Graph Harness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/harnesses": {
      "get": {
        "description": "Return registered harness templates.",
        "operationId": "list_harnesses_v1_harnesses_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessTemplateListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List harnesses",
        "tags": [
          "harnesses"
        ]
      }
    },
    "/v1/harnesses/{harness_id}": {
      "get": {
        "description": "Return one harness template by identifier.",
        "operationId": "get_harness_v1_harnesses__harness_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "harness_id",
            "required": true,
            "schema": {
              "title": "Harness Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessTemplateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one harness",
        "tags": [
          "harnesses"
        ]
      }
    },
    "/v1/spaces/{space_id}/agents/continuous-learning/runs": {
      "post": {
        "description": "Run one continuous-learning cycle and stage net-new proposals.",
        "operationId": "create_continuous_learning_run_v1_spaces__space_id__agents_continuous_learning_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContinuousLearningRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContinuousLearningRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start one harness-owned continuous-learning run",
        "tags": [
          "continuous-learning-runs"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/agents/supervisor/runs/{run_id}": {
      "get": {
        "description": "Return the persisted supervisor summary for one composed run.",
        "operationId": "get_supervisor_run_v1_spaces__space_id__agents_supervisor_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupervisorRunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one typed supervisor workflow run",
        "tags": [
          "supervisor-runs"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/runs/{run_id}": {
      "get": {
        "description": "Return one harness run from the service-local registry.",
        "operationId": "get_run_v1_spaces__space_id__runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one harness run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}